Refactored Staff Profile to adhere to MVC file structure.
