+++ components/profile/ProfileSelection.js
@@ -2,7 +2,7 @@ import React, { useLayoutEffect } from "react";
 import { useContext, useEffect, useState } from "react";
 import AuthContext from "../../context/AuthContext";
 import PatientProfileController from "./patient/PatientProfileController";
-import StaffProfile from "./staff/StaffProfile";
+import StaffProfileController from "./staff/StaffProfileController";
 import firestoreService from "../../firebase/firestoreService";
 import Staff from "../../models/Staff";
 import LoadingScreen from "../../components/LoadingScreen";
@@ -32,7 +32,7 @@ const ProfileSelection = ({ navigation, route }) => {
     if (loading) {
       return <LoadingScreen />;
     } else if (user.isStaff) {
-      return <StaffProfile navigation={navigation} route={route} />;
+      return <StaffProfileController navigation={navigation} route={route} />;
     }
     return <PatientProfileController navigation={navigation} route={route} />;
   };
