+++ components/profile/staff/StaffProfileView.js
@@ -1,4 +1,4 @@
-import React, { useState, useContext, useEffect, useLayoutEffect } from "react";
+import React from "react";
 import {
   StyleSheet,
   View,
@@ -9,8 +9,6 @@ import {
   Platform,
 } from "react-native";
 import colorDefaults from "../../../theme/colorDefaults";
-import AuthContext from "../../../context/AuthContext";
-import firestoreService from "../../../firebase/firestoreService";
 import AboutTab from "./AboutTab";
 import { Tab, TabView, Overlay } from "react-native-elements";
 import TabStyles from "../profilecomponents/TabStyles";
@@ -18,42 +16,22 @@ import GlobalProfileTab from "../profilecomponents/GlobalProfileTab";
 import LoadingScreen from "../../LoadingScreen";
 import BottomSheetNav from "../profilecomponents/BottomSheetNav";
 import UploadProfilePicture from "../profilecomponents/UploadProfilePicture";
-import { defaultPicture } from "../profilecomponents/DefaultPicture";
 import ProfilePicture from "../profilecomponents/ProfilePicture";
 
-export default function StaffProfile({ navigation, route }) {
-  // Bottom Navigation Sheet for profile picture
-  const [sheetVisible, setSheetVisible] = useState(false);
-  // Tab Index
-  const [index, setIndex] = useState(0);
-  // Passed User for search function
-  const passedUser = route.params?.user;
-  // Auth User
-  const { authUserId } = useContext(AuthContext);
-  const [user, setUser] = useState(passedUser ? passedUser : {});
-  const [loading, setLoading] = useState(passedUser ? false : true);
-  const [profilePicture, setProfilePicture] = useState(user.picture);
-  // Overlay Controls for Uploading Profile Picture
-  const [overlayVisible, setOverlayVisible] = useState(false);
-  const toggleOverlay = () => {
-    setOverlayVisible(!overlayVisible);
-  };
-
-  useLayoutEffect(() => {
-    navigation.setOptions({
-      title: "Staff Profile",
-    });
-  }, []);
-
-  !passedUser &&
-    useEffect(() => {
-      firestoreService.getUserById(authUserId).then((data) => {
-        setUser(data);
-        setProfilePicture(data.picture);
-        setLoading(false);
-      });
-    }, []);
-
+export default function StaffProfileView({
+  loading,
+  profilePicture,
+  passedUser,
+  setSheetVisible,
+  user,
+  overlayVisible,
+  toggleOverlay,
+  setProfilePicture,
+  sheetVisible,
+  index,
+  setIndex,
+  setUser,
+}) {
   const renderPage = () => {
     if (loading) {
       return <LoadingScreen />;
@@ -132,7 +110,7 @@ export default function StaffProfile({ navigation, route }) {
           </ScrollView>
         </TouchableWithoutFeedback>
         {/* This view helps Keyboard Avoiding View function properly by moving the whole display up */}
-        <View style={{ height: 100 }} />
+        <View style={Platform.OS === "ios" ? styles.ios : styles.android} />
       </KeyboardAvoidingView>
     );
   };
@@ -153,5 +131,11 @@ const styles = StyleSheet.create({
   },
   scrollViewStyle: {
     flex: 1,
+  },
+  ios: {
+    height: 100,
+  },
+  android: {
+    height: 0,
   },
 });
